+++ src/components/section-fronts/Donation/DonationLinkButton.tsx
@@ -24,8 +24,8 @@ export default function DonationLinkButton({
         fontWeight: "bold",
         // Fixes link text wrapping to 2 lines on 375 and lower mobiles.
         fontSize: {
-          base: "15px",
-          md: "initial",
+          base: "11px",
+          lg: "initial",
         },
         _hover: {
           backgroundColor: "ui.gray.light-cool",