+++ src/hooks/useDecoupledRouter.ts
@@ -3,8 +3,8 @@ import { NextRouter } from "next/router";
 const { NEXT_PUBLIC_DRUPAL_PREVIEW_SECRET } = process.env;
 
 export const DECOUPLED_ROUTER_QUERY = gql`
-  query DecoupledRouterQuery($path: String) {
-    decoupledRouter(path: $path) {
+  query DecoupledRouterQuery($path: String, $isPreview: Boolean) {
+    decoupledRouter(path: $path, isPreview: $isPreview) {
       id
       uuid
       redirect {
@@ -33,13 +33,15 @@ function useDecoupledRouter(nextRouter: NextRouter) {
   }
   // Not preview mode, so run the query to resolve a uuid from a slug.
   const slug = nextRouter.asPath;
+  // @TODO decide the order of execution, had to unable the rule because the query is running conditionally
+  // not on every render
+  // eslint-disable-next-line react-hooks/rules-of-hooks
   const { data: decoupledRouterData } = useQuery(DECOUPLED_ROUTER_QUERY, {
     variables: {
       path: slug,
     },
   });
   const uuid = decoupledRouterData?.decoupledRouter?.uuid;
-
   return {
     uuid: uuid,
     isPreview: isPreview,