+++ src/pages/research/collections/articles-databases/search.tsx
@@ -124,5 +124,4 @@ export const getServerSideProps = async () => {
   };
 };
 
-// @ts-ignore
 export default withApollo(withRedux(OnlineResourcesSearchPage));